added bones of logic to search method

+++ Campus.cpp
@@ -10,7 +10,7 @@ Campus::Campus(){
 void Campus::print_information(){
     cout << "Campus information: " << this -> campus_id << endl;
     cout << "Campus includes buildings: ";
-    for (int i = 0; i < campusBuildings.size(); i++){
+    for (int i = 0; i < (int) campusBuildings.size(); i++){
         cout << this -> campusBuildings[i].get_building_id() << ", ";
     }
 }
